pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "game_server_exporter"
version = "0.0.1"
authors = [
    {name = "nullconfig", email = "nullconfig@gmail.com"},
]
description = "exporter for valve source servers"
dependencies = [
    "python-a2s",
    "click",
    "prometheus_client",
    'importlib-metadata; python_version<"3.8"',
]
classifiers = [
    "Programming Language :: Python :: 3",
]

[project.scripts]
game-server-exporter = "game_server_exporter.interface:main"

[tool.setuptools]
packages = ["game_server_exporter"]
package-dir = {"" = "src"}
